backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
